Use rawurlencode
I'm not sure if this makes sense in all contexts. I developed an add-in share service to share the current page by email. The problem I found was that Apple's iOS email app doesn't handle `mailto:` urls with standard encoding -- all the spaces come out as `+` signs. 

Does changing the encoding to `rawurlencode` pose any problems for other services? Basically, the only difference is encoding spaces as `%20` instead of `+`...